+++ examples/custom_attribute/tests/compile-fail/syntax_error.rs
@@ -1,0 +1,4 @@
+use custom_attribute::custom_attr;
+
+#[custom_attr(foo: false)]
+mod single_attr {}