+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.6.0
+
+* Support for attribute macros
+* Update `syn`, `quote` and `proc-macro` to the newest patch
+* Update `tarpaulin` to `0.21.0`
+
 ## 0.5.0
 
 * Update syn and proc-macro dependencies